+++ jest.setup.js
@@ -1,0 +1,3 @@
+const { enterTestMode } = require('carlo')
+
+enterTestMode()