+++ src/config.ts
@@ -7,7 +7,7 @@ import osLocale from 'os-locale'
 import promisify from 'util.promisify'
 import { info, warn } from './cli'
 import { ConverterOption, ConvertType } from './converter'
-import resolveEngine, { ResolvableEngine } from './engine'
+import resolveEngine, { ResolvableEngine, ResolvedEngine } from './engine'
 import { CLIError } from './error'
 import { Theme, ThemeSet } from './theme'
 
@@ -25,6 +25,7 @@ export interface IMarpCLIArguments {
   inputDir?: string
   output?: string
   pdf?: boolean
+  preview?: boolean
   server?: boolean
   template?: string
   theme?: string
@@ -47,6 +48,7 @@ export class MarpCLIConfig {
   args: IMarpCLIArguments = {}
   conf: IMarpCLIConfig = {}
   confPath?: string
+  engine!: ResolvedEngine
 
   static moduleName = 'marp'
 
@@ -55,18 +57,21 @@ export class MarpCLIConfig {
     conf.args = args
 
     await conf.loadConf(args.configFile)
-    return conf
-  }
 
-  async converterOption(): Promise<ConverterOption> {
-    const engine = await (() => {
-      if (this.args.engine) return resolveEngine(this.args.engine)
-      if (this.conf.engine)
-        return resolveEngine(this.conf.engine, this.confPath)
+    conf.engine = await (() => {
+      if (conf.args.engine) return resolveEngine(conf.args.engine)
+      if (conf.conf.engine)
+        return resolveEngine(conf.conf.engine, conf.confPath)
 
       return resolveEngine(['@marp-team/marp-core', Marp])
     })()
 
+    return conf
+  }
+
+  private constructor() {}
+
+  async converterOption(): Promise<ConverterOption> {
     const inputDir = await this.inputDir()
     const output =
       this.args.output ||
@@ -96,6 +101,19 @@ export class MarpCLIConfig {
       initialThemes
     )
 
+    const preview = (() => {
+      const p = this.pickDefined(this.args.preview, this.conf.preview) || false
+
+      if (p && process.env.IS_DOCKER) {
+        warn(
+          `Preview window cannot show in an official docker image. Preview option was ignored.`
+        )
+        return false
+      }
+
+      return p
+    })()
+
     if (
       themeSet.themes.size <= initialThemes.length &&
       themeSetPathes.length > 0
@@ -105,6 +123,7 @@ export class MarpCLIConfig {
     return {
       inputDir,
       output,
+      preview,
       server,
       themeSet,
       allowLocalFiles:
@@ -112,12 +131,12 @@ export class MarpCLIConfig {
           this.args.allowLocalFiles,
           this.conf.allowLocalFiles
         ) || false,
-      engine: engine.klass,
+      engine: this.engine.klass,
       html: this.pickDefined(this.args.html, this.conf.html),
       lang: this.conf.lang || (await osLocale()).replace(/[_@]/g, '-'),
       options: this.conf.options || {},
-      readyScript: engine.browserScript
-        ? `<script>${engine.browserScript}</script>`
+      readyScript: this.engine.browserScript
+        ? `<script>${this.engine.browserScript}</script>`
         : undefined,
       template: this.args.template || this.conf.template || 'bespoke',
       theme: theme instanceof Theme ? theme.name : theme,
@@ -128,7 +147,10 @@ export class MarpCLIConfig {
           ? ConvertType.pdf
           : ConvertType.html,
       watch:
-        this.pickDefined(this.args.watch, this.conf.watch) || server || false,
+        this.pickDefined(this.args.watch, this.conf.watch) ||
+        preview ||
+        server ||
+        false,
     }
   }