+++ src/server.ts
@@ -8,6 +8,7 @@ import promisify from 'util.promisify'
 import { Converter, ConvertedCallback, ConvertType } from './converter'
 import { error } from './error'
 import { File, markdownExtensions } from './file'
+import favicon from './assets/favicon.png'
 import serverIndex from './server/index.pug'
 import style from './server/index.scss'
 
@@ -102,13 +103,14 @@ export class Server {
         const { name, stat } = f
         const directory = stat && stat.isDirectory()
         const parent = name === '..' && directory
+        const nodeModules = name === 'node_modules' && directory
         const convertible =
           !parent && (await this.validateMarkdown(name, stat)).valid
 
-        files.push({ convertible, directory, name, parent, stat })
+        files.push({ convertible, directory, name, nodeModules, parent, stat })
       }
 
-      callback(null, serverIndex({ directory, files, path, style }))
+      callback(null, serverIndex({ directory, favicon, files, path, style }))
     })()
   }