+++ test/_configs/custom-engine/file.js
@@ -1,0 +1,1 @@
+module.exports = { engine: './custom-engine.js' }