+++ test/_transformers/png.js
@@ -1,0 +1,6 @@
+module.exports = {
+  process: src => {
+    const uri = `data:image/png;base64,${Buffer.from(src).toString('base64')}`
+    return `module.exports = ${JSON.stringify(uri)};`
+  },
+}